float to double

+++ boarding/src/main/java/ca/ulaval/glo4002/thunderbird/boarding/domain/baggage/collection/MedicalBaggageCollection.java
@@ -36,7 +36,7 @@ public class MedicalBaggageCollection extends BaggageCollection {
     }
 
     @Override
-    public float calculateTotalCost() {
+    public double calculateTotalCost() {
         return COST;
     }
