+++ boarding/src/main/java/ca/ulaval/glo4002/thunderbird/boarding/domain/passenger/Passenger.java
@@ -78,8 +78,8 @@ public class Passenger {
         return isChild;
     }
 
-    public float calculateBaggagesPrice() {
-        float price = baggagesCollection.calculateTotalPrice();
+    public double calculateBaggagesPrice() {
+        double price = baggagesCollection.calculateTotalPrice();
         return isVip ? price * VIP_DISCOUNT : price;
     }
